backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.10.2
sqlalchemy
psycopg2-binary==2.9.10
boto3==1.35.72
python-dotenv==1.0.1
python-multipart==0.0.19
openai==1.55.3
pytest==8.3.4
pytest-cov==6.0.0
coverage==7.6.8
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
cryptography==44.0.0
pre-commit==4.0.1
pymysql==1.1.1
psycopg==3.2.3
pytest-mock==3.14.0
flake8==7.1.1
requests-mock==1.12.1